Update README description
The README description was incorrectly formed in a few areas, such as
the headings not having a space between the "#" and the text. This
formatting used to work in older versions of GitHub's markdown system,
but has since changed.

Additionally, the README description has been updated to include
the supported features of the application, as well as the new license
information.